+++ registry/doc.go
@@ -1,0 +1,3 @@
+// Package registry is a placeholder package for registry interface
+// destinations and utilities.
+package registry
