+++ registry/handlers/api_test.go
@@ -1,4 +1,4 @@
-package registry
+package handlers
 
 import (
 	"bytes"
