+++ registry/handlers/basicauth_prego14.go
@@ -1,6 +1,6 @@
 // +build !go1.4
 
-package registry
+package handlers
 
 import (
 	"encoding/base64"
