+++ registry/handlers/layer.go
@@ -1,4 +1,4 @@
-package registry
+package handlers
 
 import (
 	"net/http"
